+++ common/src/main/java/com/getwemap/example/common/SharedPreferencesExtensions.kt
@@ -1,0 +1,15 @@
+package com.getwemap.example.common
+
+import android.content.SharedPreferences
+
+interface IPreferenceKey {
+    val name: String
+}
+
+fun SharedPreferences.getString(key: IPreferenceKey): String? {
+    return getString(key.name, null)
+}
+
+fun SharedPreferences.getBoolean(key: IPreferenceKey, default: Boolean): Boolean {
+    return getBoolean(key.name, default)
+}